requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.3
astral==3.2
pytz==2024.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.2
hebrewcal==0.3.0
pyephem==4.1.3
astral==3.2
plotly==5.18.0
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
cachetools==5.3.1
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
redis==5.0.1
python-jose[cryptography]==3.3.0
